+++ Server/src/main/kotlin/ru/itmo/java/architectures/server/domain/TimedServer.kt
@@ -1,9 +1,9 @@
 package ru.itmo.java.architectures.server.domain
 
 interface TimedServer: Server {
-    val meanRequestResponseTimeMs: Long
+    val meanRequestResponseTimeMs: Double
 
-    val meanTaskTimeMs: Long
+    val meanTaskTimeMs: Double
 
     fun reset()
 }
